Retrieve user Profile
https://canvas.instructure.com/doc/api/users.html

This is basically for retrieving the user timezone. I think this is the
best way to go through this considering the current status of the
library: User and Profile share 95% of the fields and this User model
already has bio or timezone.

I'd like to know your opinion on this matter though. Thanks and regards.